Resolves #5
