+++ src/page/model.ts
@@ -1,5 +1,3 @@
-import { INode } from "./node";
-
 /**
  * Associated Unicode values for various characters.
  */
@@ -61,6 +59,66 @@ export interface IPageContent {
    * The index of the root node for the piece table for this page.
    */
   root: number;
+}
+
+/**
+ * Represents a "piece" inside the piece table.
+ */
+export interface INode {
+  /**
+   * The index of the buffer which this node refers to.
+   */
+  bufferIndex: number;
+
+  /**
+   * The start cursor for this piece, within the buffer.
+   */
+  start: IBufferCursor;
+
+  /**
+   * The end cursor for this piece, within the buffer.
+   */
+  end: IBufferCursor;
+
+  /**
+   * The count of the number of characters in the left subtree of this node.
+   */
+  leftCharCount: number;
+
+  /**
+   * The count of the number of line feeds in the left subtree of this node.
+   */
+  leftLineFeedCount: number;
+
+  /**
+   * The number of characters in this node/piece.
+   */
+  length: number;
+
+  /**
+   * The count of the number of line feeds in this node/piece.
+   */
+  lineFeedCount: number;
+
+  /**
+   * The color of this node in the tree.
+   */
+  color: Color;
+
+  /**
+   * The index of the parent to this node, in the piece table's `nodes` array.
+   */
+  parent: number;
+
+  /**
+   * The index of the left child to this node, in the piece table's `nodes` array.
+   */
+  left: number;
+
+  /**
+   * The index of the right child to this node, in the piece table's `nodes` array.
+   */
+  right: number;
 }
 
 /**
